Improved GUI
improved the GUI so that it shows
information relevent to the selected
object. All sceneObjects now have a
ofxPanel, which will hopefully be
changed to a single ofxPanel that
changes itself based on what is selected,
but this works for now. Also removed
light.color, and instead it just uses the
diffuseColor from it's material