Modified renderImage to use more tiles
renderImage no longer creates a thread for
every tile, but instead makes only as many
threads as there are cores on the CPU running
the program. Additionally, the image pieces
are now saved in a FIFO queue, and all the
threads pop off a piece, renders that piece,
and then does it again ultil there are no more
pieces left. This currently doesn't improve
performance, but will once octree is implemented